test: replace unit tests with integration tests
- Remove unit tests that duplicated code logic
- Add integration tests that import and test actual code
- Achieve 100% test coverage
- Remove mocks in favor of real server instances
